Move button to XoaButton.vue + minor fixes
